merge

+++ ArduCopter/heli_control_stabilize.pde
@@ -17,7 +17,7 @@ static bool heli_stabilize_init(bool ignore_checks)
 // should be called at 100hz or more
 static void heli_stabilize_run()
 {
-    int16_t target_roll, target_pitch;
+    float target_roll, target_pitch;
     float target_yaw_rate;
     int16_t pilot_throttle_scaled;
