+++ libraries/AP_GPS/examples/GPS_AUTO_test/GPS_AUTO_test.pde
@@ -36,6 +36,7 @@
 #include <AP_NavEKF.h>
 #include <AP_Rally.h>
 #include <AP_Scheduler.h>
+#include <AP_BattMonitor.h>
 
 const AP_HAL::HAL& hal = AP_HAL_BOARD_DRIVER;
 
@@ -72,7 +73,7 @@ void loop()
                             (int)gps.ground_course_cd() / 100,
                             gps.num_sats(),
                             gps.time_week(),
-                            gps.time_week_ms(),
+                            (unsigned long)gps.time_week_ms(),
                             gps.status());
     }
     hal.scheduler->delay(10);
